Cleaning and reorganizing the repo.
SOLPS2IMAS.jl had become longer than 1200 lines of code. Back then,
we did not practice splitting the code into multiple files. This commit
removed all unused functions and separated parsing functions in
parser.jl, SOLPS mesh index handing in solps_mesh_tools.jl, and the
subset identification functions in subset_id.jl. All tests pass.
