+++ README.md
@@ -10,12 +10,12 @@ from native SOLPS output format into IMAS
 Following code ensures it uses the enviroment described in ./Project.toml:
 
 ```
-(base) gupta@F-CJXNMY7L7 SOLPS2IMAS.jl % julia --proje
-ct=./ test/runtests.jl 
-solps2imas() time:   3.309287 seconds (9.92 M allocations: 666.357 MiB, 5.68% gc time, 71.71% compilation time)
-Test Summary: | Pass  Total  Time
-omasstuff     |    4      4  6.9s
-hello there
-Test Summary: | Pass  Total  Time
-utilities     |    3      3  0.0s
+% julia --project test/runtests.jl
+Test Summary:          | Pass  Total  Time
+Test index conversions |    1      1  0.1s
+Test Summary:       | Pass  Total  Time
+Test read_b2_output |   19     19  1.8s
+solps2imas() time:   8.995446 seconds (16.05 M allocations: 1.115 GiB, 4.49% gc time, 83.09% compilation time: <1% of which was recompilation)
+Test Summary:     | Pass  Total  Time
+Test solps2imas() |    4      4  9.9s
 ```
